# web/lib/api/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "AapatMitra API",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/reference/districts/{code}/channel-numbers": {
   "get": {
    "tags": [
     "reference"
    ],
    "summary": "Channel Numbers",
    "operationId": "channel_numbers_api_v1_reference_districts__code__channel_numbers_get",
    "parameters": [
     {
      "name": "code",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Code"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ChannelNumbersOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/villages": {
   "get": {
    "tags": [
     "reference"
    ],
    "summary": "List Villages",
    "operationId": "list_villages_api_v1_villages_get",
    "parameters": [
     {
      "name": "blockCode",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Blockcode"
      }
     },
     {
      "name": "districtCode",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Districtcode"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": {
          "type": "array",
          "items": {
           "$ref": "#/components/schemas/VillageOut"
          }
         },
         "title": "Response List Villages Api V1 Villages Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/villages/{village_id}": {
   "get": {
    "tags": [
     "reference"
    ],
    "summary": "Get Village",
    "operationId": "get_village_api_v1_villages__village_id__get",
    "parameters": [
     {
      "name": "village_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Village Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/VillageOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/facilities": {
   "get": {
    "tags": [
     "reference"
    ],
    "summary": "List Facilities",
    "operationId": "list_facilities_api_v1_facilities_get",
    "parameters": [
     {
      "name": "districtCode",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Districtcode"
      }
     },
     {
      "name": "blockCode",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Blockcode"
      }
     },
     {
      "name": "level",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Level"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": {
          "type": "array",
          "items": {
           "$ref": "#/components/schemas/FacilityOut"
          }
         },
         "title": "Response List Facilities Api V1 Facilities Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/reference/tiles/manifest": {
   "get": {
    "tags": [
     "reference"
    ],
    "summary": "Tiles Manifest",
    "description": "Offline MBTiles for one block (database.md \u00a718). Tiles are packaged by the `package_tiles` bulk job.",
    "operationId": "tiles_manifest_api_v1_reference_tiles_manifest_get",
    "parameters": [
     {
      "name": "blockCode",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "maxLength": 16,
       "title": "Blockcode"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ManifestOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/reference/lang/manifest": {
   "get": {
    "tags": [
     "reference"
    ],
    "summary": "Lang Manifest",
    "operationId": "lang_manifest_api_v1_reference_lang_manifest_get",
    "parameters": [
     {
      "name": "lang",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "pattern": "^[a-z]{2,3}$",
       "title": "Lang"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ManifestOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "ChannelNumbersOut": {
    "properties": {
     "smsNumber": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Smsnumber"
     },
     "ivrNumber": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ivrnumber"
     },
     "helpline": {
      "type": "string",
      "title": "Helpline"
     }
    },
    "type": "object",
    "required": [
     "smsNumber",
     "ivrNumber",
     "helpline"
    ],
    "title": "ChannelNumbersOut"
   },
...
   "FacilityCapabilityOut": {
    "properties": {
     "code": {
      "type": "string",
      "title": "Code"
     },
     "available": {
      "type": "boolean",
      "title": "Available"
     },
     "flaggedForReview": {
      "type": "boolean",
      "title": "Flaggedforreview"
     }
    },
    "type": "object",
    "required": [
     "code",
     "available",
     "flaggedForReview"
    ],
    "title": "FacilityCapabilityOut"
   },
...
   "FacilityOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "level": {
      "type": "string",
      "title": "Level"
     },
     "ownership": {
      "type": "string",
      "title": "Ownership"
     },
     "districtCode": {
      "type": "string",
      "title": "Districtcode"
     },
     "blockCode": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Blockcode"
     },
     "location": {
      "$ref": "#/components/schemas/GeoPointOut"
     },
     "bedsTotal": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Bedstotal"
     },
     "bedsAvailable": {
      "type": "integer",
      "title": "Bedsavailable"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "statusNote": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Statusnote"
     },
     "capabilities": {
      "items": {
       "$ref": "#/components/schemas/FacilityCapabilityOut"
      },
      "type": "array",
      "title": "Capabilities",
      "default": []
     },
     "capabilityUpdatedAt": {
      "type": "string",
      "format": "date-time",
      "title": "Capabilityupdatedat"
     },
     "stale": {
      "type": "boolean",
      "title": "Stale"
     },
     "version": {
      "type": "integer",
      "title": "Version"
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "level",
     "ownership",
     "districtCode",
     "location",
     "bedsAvailable",
     "status",
     "capabilityUpdatedAt",
     "stale",
     "version"
    ],
    "title": "FacilityOut"
   },
...
   "GeoPointOut": {
    "properties": {
     "lat": {
      "type": "number",
      "title": "Lat"
     },
     "lng": {
      "type": "number",
      "title": "Lng"
     },
     "accuracyM": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Accuracym"
     }
    },
    "type": "object",
    "required": [
     "lat",
     "lng"
    ],
    "title": "GeoPointOut"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "LinkedVillageOut": {
    "properties": {
     "villageId": {
      "type": "string",
      "format": "uuid",
      "title": "Villageid"
     },
     "searchRank": {
      "type": "integer",
      "title": "Searchrank"
     }
    },
    "type": "object",
    "required": [
     "villageId",
     "searchRank"
    ],
    "title": "LinkedVillageOut"
   },
...
   "ManifestOut": {
    "properties": {
     "version": {
      "type": "integer",
      "title": "Version"
     },
     "url": {
      "type": "string",
      "title": "Url"
     },
     "sha256": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sha256"
     },
     "sizeBytes": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sizebytes"
     }
    },
    "type": "object",
    "required": [
     "version",
     "url"
    ],
    "title": "ManifestOut"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
   "VillageOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "blockCode": {
      "type": "string",
      "title": "Blockcode"
     },
     "districtCode": {
      "type": "string",
      "title": "Districtcode"
     },
     "location": {
      "$ref": "#/components/schemas/GeoPointOut"
     },
     "waypoints": {
      "items": {
       "$ref": "#/components/schemas/WaypointOut"
      },
      "type": "array",
      "title": "Waypoints",
      "default": []
     },
     "linkedVillages": {
      "items": {
       "$ref": "#/components/schemas/LinkedVillageOut"
      },
      "type": "array",
      "title": "Linkedvillages",
      "default": []
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "blockCode",
     "districtCode",
     "location"
    ],
    "title": "VillageOut"
   },
...
   "WaypointOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "kind": {
      "type": "string",
      "title": "Kind"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "point": {
      "$ref": "#/components/schemas/GeoPointOut"
     },
     "isDefault": {
      "type": "boolean",
      "title": "Isdefault"
     },
     "vehicleKindNeeded": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Vehiclekindneeded"
     }
    },
    "type": "object",
    "required": [
     "id",
     "kind",
     "name",
     "point",
     "isDefault"
    ],
    "title": "WaypointOut"
   }
  }
 }
}